[MINOR] Remove warinnings and clear solution

+++ C++/ClustDesign/odeint.h
@@ -14,7 +14,7 @@ using namespace boost::numeric::odeint;
 /*** TODO Change hard coding !!!***/
 const int state_value = 2;
 
-typedef float x_type;
+typedef double x_type;
 typedef boost::array< x_type, state_value > state_type;
 typedef vector<x_type> parameters_type;
 typedef boost::numeric::ublas::matrix<x_type> matrix_type;
